+++ packages/ai/core/generate-text/generate-text.ts
@@ -6,6 +6,7 @@ import { createIdGenerator, IdGenerator } from '@ai-sdk/provider-utils';
 import { Tracer } from '@opentelemetry/api';
 import { NoOutputSpecifiedError } from '../../src/error/no-output-specified-error';
 import { ToolExecutionError } from '../../src/error/tool-execution-error';
+import { asArray } from '../../src/util/as-array';
 import { prepareRetries } from '../../src/util/prepare-retries';
 import { ModelMessage } from '../prompt';
 import { CallSettings } from '../prompt/call-settings';
@@ -29,7 +30,11 @@ import { Output } from './output';
 import { parseToolCall } from './parse-tool-call';
 import { ResponseMessage } from './response-message';
 import { DefaultStepResult, StepResult } from './step-result';
-import { maxSteps, StopCondition } from './stop-condition';
+import {
+  isStopConditionMet,
+  stepCountIs,
+  StopCondition,
+} from './stop-condition';
 import { toResponseMessages } from './to-response-messages';
 import { ToolCallArray } from './tool-call';
 import { ToolCallRepairFunction } from './tool-call-repair';
@@ -111,12 +116,14 @@ export async function generateText<
   maxRetries: maxRetriesArg,
   abortSignal,
   headers,
-  continueUntil = maxSteps(1),
+  stopWhen = stepCountIs(1),
   experimental_output: output,
   experimental_telemetry: telemetry,
   providerOptions,
-  experimental_activeTools: activeTools,
-  experimental_prepareStep: prepareStep,
+  experimental_activeTools,
+  activeTools = experimental_activeTools,
+  experimental_prepareStep,
+  prepareStep = experimental_prepareStep,
   experimental_repairToolCall: repairToolCall,
   _internal: {
     generateId = originalGenerateId,
@@ -141,7 +148,15 @@ The tool choice strategy. Default: 'auto'.
      */
     toolChoice?: ToolChoice<NoInfer<TOOLS>>;
 
-    continueUntil?: StopCondition<NoInfer<TOOLS>>;
+    /**
+Condition for stopping the generation when there are tool results in the last step.
+When the condition is an array, any of the conditions can be met to stop the generation.
+
+@default stepCountIs(1)
+     */
+    stopWhen?:
+      | StopCondition<NoInfer<TOOLS>>
+      | Array<StopCondition<NoInfer<TOOLS>>>;
 
     /**
 Optional telemetry configuration (experimental).
@@ -156,15 +171,36 @@ functionality that can be fully encapsulated in the provider.
     providerOptions?: ProviderOptions;
 
     /**
+     * @deprecated Use `activeTools` instead.
+     */
+    experimental_activeTools?: Array<keyof NoInfer<TOOLS>>;
+
+    /**
 Limits the tools that are available for the model to call without
 changing the tool call and result types in the result.
      */
-    experimental_activeTools?: Array<keyof NoInfer<TOOLS>>;
+    activeTools?: Array<keyof NoInfer<TOOLS>>;
 
     /**
 Optional specification for parsing structured outputs from the LLM response.
      */
     experimental_output?: Output<OUTPUT, OUTPUT_PARTIAL>;
+
+    /**
+     * @deprecated Use `prepareStep` instead.
+     */
+    experimental_prepareStep?: (options: {
+      steps: Array<StepResult<NoInfer<TOOLS>>>;
+      stepNumber: number;
+      model: LanguageModel;
+    }) => PromiseLike<
+      | {
+          model?: LanguageModel;
+          toolChoice?: ToolChoice<NoInfer<TOOLS>>;
+          activeTools?: Array<keyof NoInfer<TOOLS>>;
+        }
+      | undefined
+    >;
 
     /**
 Optional function that you can use to provide different settings for a step.
@@ -177,7 +213,7 @@ Optional function that you can use to provide different settings for a step.
 @returns An object that contains the settings for the step.
 If you return undefined (or for undefined settings), the settings from the outer level will be used.
     */
-    experimental_prepareStep?: (options: {
+    prepareStep?: (options: {
       steps: Array<StepResult<NoInfer<TOOLS>>>;
       stepNumber: number;
       model: LanguageModel;
@@ -185,7 +221,7 @@ If you return undefined (or for undefined settings), the settings from the outer
       | {
           model?: LanguageModel;
           toolChoice?: ToolChoice<NoInfer<TOOLS>>;
-          experimental_activeTools?: Array<keyof NoInfer<TOOLS>>;
+          activeTools?: Array<keyof NoInfer<TOOLS>>;
         }
       | undefined
     >;
@@ -208,6 +244,7 @@ A function that attempts to repair a tool call that failed to parse.
       currentDate?: () => Date;
     };
   }): Promise<GenerateTextResult<TOOLS, OUTPUT>> {
+  const stopConditions = asArray(stopWhen);
   const { maxRetries, retry } = prepareRetries({ maxRetries: maxRetriesArg });
 
   const callSettings = prepareCallSettings(settings);
@@ -283,8 +320,7 @@ A function that attempts to repair a tool call that failed to parse.
           prepareToolsAndToolChoice({
             tools,
             toolChoice: prepareStepResult?.toolChoice ?? toolChoice,
-            activeTools:
-              prepareStepResult?.experimental_activeTools ?? activeTools,
+            activeTools: prepareStepResult?.activeTools ?? activeTools,
           });
 
         currentModelResponse = await retry(() =>
@@ -459,8 +495,8 @@ A function that attempts to repair a tool call that failed to parse.
         currentToolCalls.length > 0 &&
         // all current tool calls have results:
         currentToolResults.length === currentToolCalls.length &&
-        // continue until the stop condition is met:
-        !(await continueUntil({ steps }))
+        // continue until a stop condition is met:
+        !(await isStopConditionMet({ stopConditions, steps }))
       );
 
       // Add response information to the span: