+++ packages/ai/core/generate-text/stream-text.ts
@@ -65,6 +65,11 @@ import { ToolCallRepairFunction } from './tool-call-repair';
 import { ToolResultUnion } from './tool-result';
 import { ToolSet } from './tool-set';
 import { stringifyForTelemetry } from '../prompt/stringify-for-telemetry';
+import {
+  isStopConditionMet,
+  stepCountIs,
+  StopCondition,
+} from './stop-condition';
 
 const originalGenerateId = createIdGenerator({
   prefix: 'aitxt',
@@ -202,13 +207,15 @@ export function streamText<
   maxRetries,
   abortSignal,
   headers,
-  maxSteps = 1,
+  stopWhen = stepCountIs(1),
   experimental_output: output,
   experimental_telemetry: telemetry,
+  prepareStep,
   providerOptions,
   experimental_toolCallStreaming = false,
   toolCallStreaming = experimental_toolCallStreaming,
-  experimental_activeTools: activeTools,
+  experimental_activeTools,
+  activeTools = experimental_activeTools,
   experimental_repairToolCall: repairToolCall,
   experimental_transform: transform,
   onChunk,
@@ -239,13 +246,14 @@ The tool choice strategy. Default: 'auto'.
     toolChoice?: ToolChoice<TOOLS>;
 
     /**
-Maximum number of sequential LLM calls (steps), e.g. when you use tool calls. Must be at least 1.
+Condition for stopping the generation when there are tool results in the last step.
+When the condition is an array, any of the conditions can be met to stop the generation.
 
-A maximum number is required to prevent infinite loops in the case of misconfigured tools.
-
-By default, it's set to 1, which means that only a single LLM call is made.
- */
-    maxSteps?: number;
+@default stepCountIs(1)
+     */
+    stopWhen?:
+      | StopCondition<NoInfer<TOOLS>>
+      | Array<StopCondition<NoInfer<TOOLS>>>;
 
     /**
 Optional telemetry configuration (experimental).
@@ -260,15 +268,44 @@ functionality that can be fully encapsulated in the provider.
     providerOptions?: ProviderOptions;
 
     /**
-Limits the tools that are available for the model to call without
-changing the tool call and result types in the result.
+     * @deprecated Use `activeTools` instead.
      */
-    experimental_activeTools?: Array<keyof TOOLS>;
+    experimental_activeTools?: Array<keyof NoInfer<TOOLS>>;
+
+    /**
+   Limits the tools that are available for the model to call without
+   changing the tool call and result types in the result.
+        */
+    activeTools?: Array<keyof NoInfer<TOOLS>>;
 
     /**
 Optional specification for parsing structured outputs from the LLM response.
      */
     experimental_output?: Output<OUTPUT, PARTIAL_OUTPUT>;
+
+    /**
+Optional function that you can use to provide different settings for a step.
+
+@param options - The options for the step.
+@param options.steps - The steps that have been executed so far.
+@param options.stepNumber - The number of the step that is being executed.
+@param options.model - The model that is being used.
+
+@returns An object that contains the settings for the step.
+If you return undefined (or for undefined settings), the settings from the outer level will be used.
+    */
+    prepareStep?: (options: {
+      steps: Array<StepResult<NoInfer<TOOLS>>>;
+      stepNumber: number;
+      model: LanguageModel;
+    }) => PromiseLike<
+      | {
+          model?: LanguageModel;
+          toolChoice?: ToolChoice<NoInfer<TOOLS>>;
+          activeTools?: Array<keyof NoInfer<TOOLS>>;
+        }
+      | undefined
+    >;
 
     /**
 A function that attempts to repair a tool call that failed to parse.
@@ -345,9 +382,10 @@ Internal. For test use only. May change without notice.
     transforms: asArray(transform),
     activeTools,
     repairToolCall,
-    maxSteps,
+    stopConditions: asArray(stopWhen),
     output,
     providerOptions,
+    prepareStep,
     onChunk,
     onError,
     onFinish,
@@ -446,13 +484,13 @@ function createOutputTransformStream<
 class DefaultStreamTextResult<TOOLS extends ToolSet, OUTPUT, PARTIAL_OUTPUT>
   implements StreamTextResult<TOOLS, PARTIAL_OUTPUT>
 {
-  private readonly totalUsagePromise = new DelayedPromise<
+  private readonly _totalUsage = new DelayedPromise<
     Awaited<StreamTextResult<TOOLS, PARTIAL_OUTPUT>['usage']>
   >();
-  private readonly finishReasonPromise = new DelayedPromise<
+  private readonly _finishReason = new DelayedPromise<
     Awaited<StreamTextResult<TOOLS, PARTIAL_OUTPUT>['finishReason']>
   >();
-  private readonly stepsPromise = new DelayedPromise<
+  private readonly _steps = new DelayedPromise<
     Awaited<StreamTextResult<TOOLS, PARTIAL_OUTPUT>['steps']>
   >();
 
@@ -484,9 +522,10 @@ class DefaultStreamTextResult<TOOLS extends ToolSet, OUTPUT, PARTIAL_OUTPUT>
     transforms,
     activeTools,
     repairToolCall,
-    maxSteps,
+    stopConditions,
     output,
     providerOptions,
+    prepareStep,
     now,
     currentDate,
     generateId,
@@ -510,9 +549,23 @@ class DefaultStreamTextResult<TOOLS extends ToolSet, OUTPUT, PARTIAL_OUTPUT>
     transforms: Array<StreamTextTransform<TOOLS>>;
     activeTools: Array<keyof TOOLS> | undefined;
     repairToolCall: ToolCallRepairFunction<TOOLS> | undefined;
-    maxSteps: number;
+    stopConditions: Array<StopCondition<NoInfer<TOOLS>>>;
     output: Output<OUTPUT, PARTIAL_OUTPUT> | undefined;
     providerOptions: ProviderOptions | undefined;
+    prepareStep:
+      | ((options: {
+          steps: Array<StepResult<NoInfer<TOOLS>>>;
+          stepNumber: number;
+          model: LanguageModel;
+        }) => PromiseLike<
+          | {
+              model?: LanguageModel;
+              toolChoice?: ToolChoice<NoInfer<TOOLS>>;
+              activeTools?: Array<keyof NoInfer<TOOLS>>;
+            }
+          | undefined
+        >)
+      | undefined;
     now: () => number;
     currentDate: () => Date;
     generateId: () => string;
@@ -523,16 +576,13 @@ class DefaultStreamTextResult<TOOLS extends ToolSet, OUTPUT, PARTIAL_OUTPUT>
     onFinish: undefined | StreamTextOnFinishCallback<TOOLS>;
     onStepFinish: undefined | StreamTextOnStepFinishCallback<TOOLS>;
   }) {
-    if (maxSteps < 1) {
-      throw new InvalidArgumentError({
-        parameter: 'maxSteps',
-        value: maxSteps,
-        message: 'maxSteps must be at least 1',
-      });
-    }
-
     this.output = output;
     this.generateId = generateId;
+
+    // promise to ensure that the step has been fully processed by the event processor
+    // before a new step is started. This is required because the continuation condition
+    // needs the updated steps to determine if another step is needed.
+    let stepFinish!: DelayedPromise<void>;
 
     let activeReasoningPart:
       | undefined
@@ -652,6 +702,10 @@ class DefaultStreamTextResult<TOOLS extends ToolSet, OUTPUT, PARTIAL_OUTPUT>
           activeReasoningPart = undefined;
 
           recordedResponseMessages.push(...stepMessages);
+
+          // resolve the promise to signal that the step has been fully processed
+          // by the event processor:
+          stepFinish.resolve();
         }
 
         if (part.type === 'finish') {
@@ -675,11 +729,11 @@ class DefaultStreamTextResult<TOOLS extends ToolSet, OUTPUT, PARTIAL_OUTPUT>
           };
 
           // from finish:
-          self.finishReasonPromise.resolve(finishReason);
-          self.totalUsagePromise.resolve(totalUsage);
+          self._finishReason.resolve(finishReason);
+          self._totalUsage.resolve(totalUsage);
 
           // aggregate results:
-          self.stepsPromise.resolve(recordedSteps);
+          self._steps.resolve(recordedSteps);
 
           // call onFinish callback:
           const finalStep = recordedSteps[recordedSteps.length - 1];
@@ -793,7 +847,6 @@ class DefaultStreamTextResult<TOOLS extends ToolSet, OUTPUT, PARTIAL_OUTPUT>
           'ai.prompt': {
             input: () => JSON.stringify({ system, prompt, messages }),
           },
-          'ai.settings.maxSteps': maxSteps,
         },
       }),
       tracer,
@@ -810,6 +863,8 @@ class DefaultStreamTextResult<TOOLS extends ToolSet, OUTPUT, PARTIAL_OUTPUT>
           responseMessages: Array<ResponseMessage>;
           usage: LanguageModelUsage;
         }) {
+          stepFinish = new DelayedPromise<void>();
+
           const initialPrompt = await standardizePrompt({
             system,
             prompt,
@@ -821,6 +876,12 @@ class DefaultStreamTextResult<TOOLS extends ToolSet, OUTPUT, PARTIAL_OUTPUT>
             ...responseMessages,
           ];
 
+          const prepareStepResult = await prepareStep?.({
+            model,
+            steps: recordedSteps,
+            stepNumber: recordedSteps.length,
+          });
+
           const promptMessages = await convertToLanguageModelPrompt({
             prompt: {
               system: initialPrompt.system,
@@ -829,9 +890,13 @@ class DefaultStreamTextResult<TOOLS extends ToolSet, OUTPUT, PARTIAL_OUTPUT>
             supportedUrls: await model.supportedUrls,
           });
 
-          const toolsAndToolChoice = {
-            ...prepareToolsAndToolChoice({ tools, toolChoice, activeTools }),
-          };
+          const stepModel = prepareStepResult?.model ?? model;
+          const { toolChoice: stepToolChoice, tools: stepTools } =
+            prepareToolsAndToolChoice({
+              tools,
+              toolChoice: prepareStepResult?.toolChoice ?? toolChoice,
+              activeTools: prepareStepResult?.activeTools ?? activeTools,
+            });
 
           const {
             result: { stream, response, request },
@@ -848,26 +913,27 @@ class DefaultStreamTextResult<TOOLS extends ToolSet, OUTPUT, PARTIAL_OUTPUT>
                     telemetry,
                   }),
                   ...baseTelemetryAttributes,
+                  // model:
+                  'ai.model.provider': stepModel.provider,
+                  'ai.model.id': stepModel.modelId,
+                  // prompt:
                   'ai.prompt.messages': {
                     input: () => stringifyForTelemetry(promptMessages),
                   },
                   'ai.prompt.tools': {
                     // convert the language model level tools:
-                    input: () =>
-                      toolsAndToolChoice.tools?.map(tool =>
-                        JSON.stringify(tool),
-                      ),
+                    input: () => stepTools?.map(tool => JSON.stringify(tool)),
                   },
                   'ai.prompt.toolChoice': {
                     input: () =>
-                      toolsAndToolChoice.toolChoice != null
-                        ? JSON.stringify(toolsAndToolChoice.toolChoice)
+                      stepToolChoice != null
+                        ? JSON.stringify(stepToolChoice)
                         : undefined,
                   },
 
                   // standardized gen-ai llm span attributes:
-                  'gen_ai.system': model.provider,
-                  'gen_ai.request.model': model.modelId,
+                  'gen_ai.system': stepModel.provider,
+                  'gen_ai.request.model': stepModel.modelId,
                   'gen_ai.request.frequency_penalty':
                     callSettings.frequencyPenalty,
                   'gen_ai.request.max_tokens': callSettings.maxOutputTokens,
@@ -885,9 +951,10 @@ class DefaultStreamTextResult<TOOLS extends ToolSet, OUTPUT, PARTIAL_OUTPUT>
                 return {
                   startTimestampMs: now(), // get before the call
                   doStreamSpan,
-                  result: await model.doStream({
+                  result: await stepModel.doStream({
                     ...callSettings,
-                    ...toolsAndToolChoice,
+                    tools: stepTools,
+                    toolChoice: stepToolChoice,
                     responseFormat: output?.responseFormat,
                     prompt: promptMessages,
                     providerOptions,
@@ -899,7 +966,7 @@ class DefaultStreamTextResult<TOOLS extends ToolSet, OUTPUT, PARTIAL_OUTPUT>
             }),
           );
 
-          const transformedStream = runToolsTransformation({
+          const streamWithToolResults = runToolsTransformation({
             tools,
             generatorStream: stream,
             toolCallStreaming,
@@ -949,7 +1016,7 @@ class DefaultStreamTextResult<TOOLS extends ToolSet, OUTPUT, PARTIAL_OUTPUT>
           }
 
           self.addStream(
-            transformedStream.pipeThrough(
+            streamWithToolResults.pipeThrough(
               new TransformStream<
                 SingleRequestTextStreamPart<TOOLS>,
                 TextStreamPart<TOOLS>
@@ -1155,11 +1222,19 @@ class DefaultStreamTextResult<TOOLS extends ToolSet, OUTPUT, PARTIAL_OUTPUT>
 
                   const combinedUsage = addLanguageModelUsage(usage, stepUsage);
 
+                  // wait for the step to be fully processed by the event processor
+                  // to ensure that the recorded steps are complete:
+                  await stepFinish.promise;
+
                   if (
-                    currentStep + 1 < maxSteps && // there are tool calls:
                     stepToolCalls.length > 0 &&
                     // all current tool calls have results:
-                    stepToolResults.length === stepToolCalls.length
+                    stepToolResults.length === stepToolCalls.length &&
+                    // continue until a stop condition is met:
+                    !(await isStopConditionMet({
+                      stopConditions,
+                      steps: recordedSteps,
+                    }))
                   ) {
                     // append to messages for the next step:
                     responseMessages.push(
@@ -1215,7 +1290,7 @@ class DefaultStreamTextResult<TOOLS extends ToolSet, OUTPUT, PARTIAL_OUTPUT>
   }
 
   get steps() {
-    return this.stepsPromise.value;
+    return this._steps.promise;
   }
 
   private get finalStep() {
@@ -1275,11 +1350,11 @@ class DefaultStreamTextResult<TOOLS extends ToolSet, OUTPUT, PARTIAL_OUTPUT>
   }
 
   get totalUsage() {
-    return this.totalUsagePromise.value;
+    return this._totalUsage.promise;
   }
 
   get finishReason() {
-    return this.finishReasonPromise.value;
+    return this._finishReason.promise;
   }
 
   /**
@@ -1415,9 +1490,8 @@ However, the LLM results are expected to be small enough to not cause issues.
             case 'source': {
               if (sendSources) {
                 controller.enqueue({
-                  type: 'source',
-                  sourceType: part.sourceType,
-                  id: part.id,
+                  type: 'source-url',
+                  sourceId: part.id,
                   url: part.url,
                   title: part.title,
                   providerMetadata: part.providerMetadata,