+++ packages/provider/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @ai-sdk/provider
 
+## 2.0.0-alpha.4
+
+### Patch Changes
+
+- dc714f3: release alpha.4
+
+## 2.0.0-alpha.3
+
+### Patch Changes
+
+- 6b98118: release alpha.3
+
+## 2.0.0-alpha.2
+
+### Patch Changes
+
+- 26535e0: release alpha.2
+
 ## 2.0.0-alpha.1
 
 ### Patch Changes