+++ packages/rsc/CHANGELOG.md
@@ -1,5 +1,43 @@
 # @ai-sdk/rsc
 
+## 1.0.0-alpha.4
+
+### Patch Changes
+
+- Updated dependencies [b32c141]
+- Updated dependencies [dc714f3]
+- Updated dependencies [72d7d72]
+- Updated dependencies [9315076]
+- Updated dependencies [7d97ab6]
+- Updated dependencies [37a916d]
+- Updated dependencies [5f2b3d4]
+  - ai@5.0.0-alpha.4
+  - @ai-sdk/provider@2.0.0-alpha.4
+  - @ai-sdk/provider-utils@3.0.0-alpha.4
+
+## 1.0.0-alpha.3
+
+### Patch Changes
+
+- Updated dependencies [ab7ccef]
+- Updated dependencies [257224b]
+- Updated dependencies [6b98118]
+- Updated dependencies [0463011]
+- Updated dependencies [d306260]
+  - ai@5.0.0-alpha.3
+  - @ai-sdk/provider@2.0.0-alpha.3
+  - @ai-sdk/provider-utils@3.0.0-alpha.3
+
+## 1.0.0-alpha.2
+
+### Patch Changes
+
+- Updated dependencies [82aa95d]
+- Updated dependencies [26535e0]
+  - ai@5.0.0-alpha.2
+  - @ai-sdk/provider@2.0.0-alpha.2
+  - @ai-sdk/provider-utils@3.0.0-alpha.2
+
 ## 1.0.0-alpha.1
 
 ### Patch Changes