+++ packages/valibot/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @ai-sdk/valibot
 
+## 1.0.0-alpha.4
+
+### Patch Changes
+
+- @ai-sdk/provider-utils@3.0.0-alpha.4
+
+## 1.0.0-alpha.3
+
+### Patch Changes
+
+- @ai-sdk/provider-utils@3.0.0-alpha.3
+
+## 1.0.0-alpha.2
+
+### Patch Changes
+
+- @ai-sdk/provider-utils@3.0.0-alpha.2
+
 ## 1.0.0-alpha.1
 
 ### Patch Changes